// packages/ioc/src/interface/decorators/options/class-options.interface.ts
export interface ClassOptions {
  /**
   * The class uuid.
   * @type {string}
   * @memberof ClassMetadata
   */
  uuid?: string;
  /**
   * The name of the class.
   * @type {string}
   * @memberof ClassMetadata
   */
  name?: string;
  /**
   * The original name of the class.
   * @type {string}
   * @memberof ClassMetadata
   */
  originalName?: string;
}